enableReducer, move interface to reducer
