+++ packages/suite/src/actions/suite/__tests__/firmwareActions.test.ts
@@ -75,7 +75,7 @@ export const getInitialState = (override: any) => {
             locks: [3],
         },
         firmware: {
-            initialized: true,
+            reducerEnabled: true,
             status: null,
         },
     };
