+++ packages/suite/src/utils/wallet/__tests__/networkUtils.test.ts
@@ -3,7 +3,7 @@ import { httpRequest } from '../networkUtils';
 const setMock = (mock: any) => {
     // @ts-ignore
     global.fetch = jest.fn().mockImplementation(() => {
-        const p = new Promise((resolve) => {
+        const p = new Promise(resolve => {
             resolve({
                 ok: mock.ok,
                 json: () =>
